@@ -1,21 +1,21 @@
 ActiveAdmin.register AdminUser do
-  menu :priority => 2     
-  index do                            
-    column :email                     
-    column :current_sign_in_at        
-    column :last_sign_in_at           
-    column :sign_in_count             
-    default_actions                 
-  end                                 
+  menu :priority => 2
+  index do
+    column :email
+    column :current_sign_in_at
+    column :last_sign_in_at
+    column :sign_in_count
+    actions
+  end
 
-  config.filters= false                      
+  config.filters= false
 
-  form do |f|                         
-    f.inputs "Admin Details" do       
-      f.input :email                  
-      f.input :password               
-      f.input :password_confirmation  
-    end                               
-    f.actions                         
-  end                                 
-end                                   
+  form do |f|
+    f.inputs "Admin Details" do
+      f.input :email
+      f.input :password
+      f.input :password_confirmation
+    end
+    f.actions
+  end
+end
